Added a reader macro to be able to emit xml tags without defining them. Use #<(tag :attr value #<(tag2)) form or lisp code as body. Nesting is supported just like with predefined tags.
Actually it is possible to support #<alma :attr value > syntax, but it requires the extra space before the closing sign which is error prone. I could not figure a way to workaround this. The > character is not a special char, so it may be part of normal symbols.

darcs-hash:20060319131650-e3083-2094ef60c2020123de74ac9da2c8bee9e804d4a3